backend/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Flask==2.3.2
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
supabase>=2.0.0
